vit/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
tqdm>=4.60.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pyyaml>=6.0
optuna>=3.0.0
Pillow>=9.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6